tadaaaa !

+++ ft_max.c
@@ -6,15 +6,15 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 13:20:55 by vviterbo          #+#    #+#             */
-/*   Updated: 2024/11/05 15:40:39 by vviterbo         ###   ########.fr       */
+/*   Updated: 2024/11/06 10:26:09 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_max(long a, long b);
+float	ft_max(float a, float b);
 
-long	ft_max(long a, long b)
+float	ft_max(float a, float b)
 {
 	if (a >= b)
 		return (a);
