+++ ft_min.c
@@ -6,15 +6,15 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 13:19:04 by vviterbo          #+#    #+#             */
-/*   Updated: 2024/11/05 15:40:36 by vviterbo         ###   ########.fr       */
+/*   Updated: 2024/11/06 10:25:59 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_min(long a, long b);
+float	ft_min(float a, float b);
 
-long	ft_min(long a, long b)
+float	ft_min(float a, float b)
 {
 	if (a <= b)
 		return (a);
